+++ axelor-supplychain/src/main/java/com/axelor/apps/supplychain/service/declarationofexchanges/DeclarationOfExchangesExporterGoods.java
@@ -168,7 +168,7 @@ public class DeclarationOfExchangesExporterGoods extends DeclarationOfExchangesE
 
     BigDecimal fiscalValue =
         stockMoveLine
-            .getCompanyUnitPriceUntaxed()
+            .getValuatedUnitPrice()
             .multiply(stockMoveLine.getRealQty())
             .setScale(0, RoundingMode.HALF_UP);